requirements.txt
allure-pytest==2.15.3
pytest==8.3.3
selenium==4.36.0
